# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/households": {
      "post": {
        "summary": "Create Household",
        "operationId": "create_household_households_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HouseholdCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/households/join": {
      "post": {
        "summary": "Join Household",
        "operationId": "join_household_households_join_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HouseholdJoin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/consumables": {
      "post": {
        "summary": "Create Consumable",
        "operationId": "create_consumable_consumables_post",
        "parameters": [
          {
            "name": "x-household-code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Household-Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumableCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Consumables",
        "operationId": "list_consumables_consumables_get",
        "parameters": [
          {
            "name": "x-household-code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Household-Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/consumables/{id}": {
      "get": {
        "summary": "Get Consumable",
        "operationId": "get_consumable_consumables__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-household-code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Household-Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Consumable",
        "operationId": "update_consumable_consumables__id__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-household-code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Household-Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumableUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Consumable",
        "operationId": "delete_consumable_consumables__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-household-code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Household-Code"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/consumables/{id}/status": {
      "put": {
        "summary": "Set Consumable Status",
        "operationId": "set_consumable_status_consumables__id__status_put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-household-code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Household-Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/consumables/{id}/image": {
      "put": {
        "summary": "Set Consumable Image",
        "operationId": "set_consumable_image_consumables__id__image_put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-household-code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Household-Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "image/jpeg": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "post": {
        "summary": "Create Task",
        "operationId": "create_task_tasks_post",
        "parameters": [
          {
            "name": "x-household-code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Household-Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Tasks",
        "operationId": "list_tasks_tasks_get",
        "parameters": [
          {
            "name": "x-household-code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Household-Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}/done": {
      "post": {
        "summary": "Complete Task",
        "operationId": "complete_task_tasks__id__done_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-household-code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Household-Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConsumableCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "status": {
            "$ref": "#/components/schemas/Status",
            "default": "full"
          }
        },
        "type": "object",
        "required": [
          "name",
          "period"
        ],
        "title": "ConsumableCreate"
      },
      "ConsumableUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        },
        "type": "object",
        "required": [
          "name",
          "period",
          "status"
        ],
        "title": "ConsumableUpdate"
      },
      "Difficulty": {
        "type": "string",
        "enum": [
          "light",
          "medium",
          "heavy"
        ],
        "title": "Difficulty"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HouseholdCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "HouseholdCreate"
      },
      "HouseholdJoin": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "HouseholdJoin"
      },
      "HouseholdRead": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "HouseholdRead"
      },
      "Period": {
        "type": "string",
        "enum": [
          "weekly",
          "biweekly",
          "monthly",
          "every 2 months",
          "every 3 months",
          "every 4 months",
          "every 5 months",
          "every 6 months",
          "every 7 months",
          "every 8 months",
          "every 9 months",
          "every 10 months",
          "every 11 months",
          "yearly"
        ],
        "title": "Period"
      },
      "Status": {
        "type": "string",
        "enum": [
          "full",
          "coming to an end",
          "finished"
        ],
        "title": "Status"
      },
      "StatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StatusUpdate"
      },
      "TaskCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty",
            "default": "medium"
          }
        },
        "type": "object",
        "required": [
          "name",
          "period"
        ],
        "title": "TaskCreate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}